not deleting tmp dir

+++ scripts/cli/src/commands/cleanup-cloud-resources.sh
@@ -3,7 +3,12 @@ force="${args[--force]}"
 
 set +e
 tmp_dir=$(mktemp -d -t pg-XXXXXXXXXX)
-trap 'rm -rf $tmp_dir' EXIT
+if [ -z "$PG_VERBOSE_MODE" ]
+then
+    trap 'rm -rf $tmp_dir' EXIT
+else
+    log "🐛📂 not deleting tmp dir $tmp_dir"
+fi
 
 if [[ ! -n "$user" ]]
 then
