+++ scripts/cli/src/commands/connector/show-config-parameters.sh
@@ -25,7 +25,12 @@ then
 fi
 
 tmp_dir=$(mktemp -d -t pg-XXXXXXXXXX)
-trap 'rm -rf $tmp_dir' EXIT
+if [ -z "$PG_VERBOSE_MODE" ]
+then
+    trap 'rm -rf $tmp_dir' EXIT
+else
+    log "🐛📂 not deleting tmp dir $tmp_dir"
+fi
 json_file=$tmp_dir/connector.json
 
 items=($connector)
