+++ scripts/cli/src/commands/connector/show-lag.sh
@@ -135,7 +135,12 @@ function log_same() {
 }
 
 tmp_dir=$(mktemp -d -t pg-XXXXXXXXXX)
-trap 'rm -rf $tmp_dir' EXIT
+if [ -z "$PG_VERBOSE_MODE" ]
+then
+    trap 'rm -rf $tmp_dir' EXIT
+else
+    log "🐛📂 not deleting tmp dir $tmp_dir"
+fi
 lag_output=$tmp_dir/lag_output
 
 function handle_signal {
