+++ scripts/cli/src/commands/connector/snippets.sh
@@ -2,7 +2,12 @@ converter=${args[--converter]}
 dlq=${args[--dlq]}
 
 tmp_dir=$(mktemp -d -t pg-XXXXXXXXXX)
-trap 'rm -rf $tmp_dir' EXIT
+if [ -z "$PG_VERBOSE_MODE" ]
+then
+    trap 'rm -rf $tmp_dir' EXIT
+else
+    log "🐛📂 not deleting tmp dir $tmp_dir"
+fi
 
 if [[ ! -n "$converter" ]] && [[ ! -n "$dlq" ]]
 then
